+++ R/data_plod_pred.R
@@ -1,9 +1,13 @@
 #' Species climate optimum dataset.
 #'
 #' Climate defined per species from Kunstler et al. 2021. Each species
-#' had its IPM integrated for 3 climates : optimum on climate 2,
-#' hot edge at climate 1 and cold edge for climate 3. Climate was defined
-#' from sgdd and wai variables.
+#' had its IPM integrated for 3 climates :
+#' \describe{
+#' \item{1}{hot edge}
+#' \item{2}{optimum climate for the species}
+#' \itme{3}{cold edge}
+#'
+#' Climate was defined from sgdd and wai variables.
 #'
 #' @name climate_species
 #'