+++ dev/run_all_test.R
@@ -4,7 +4,7 @@ devtools::load_all('.')
 devtools::document('.')
 # attachment::att_amend_desc()
 devtools::test()
-devtools::check(vignettes = FALSE)
+devtools::check()
 
 
 # change name